shortened line in triangle.c

+++ triangle.c
@@ -46,8 +46,9 @@ void draw_triangle(triangle_t *this)
 		swap(&this->v1, &this->v2);
 
 	// v1.y <= v2.y <= v3.y
-	uint16_t newX = (uint16_t) (this->v1.x +
-	((float)(this->v2.y - this->v1.y) / (float)(this->v3.y - this->v1.y)) * (this->v3.x - this->v1.x));
+	uint16_t newX = (this->v1.x + (float)
+	(this->v2.y - this->v1.y) / (this->v3.y - this->v1.y) *
+	(this->v3.x - this->v1.x));
 	fillTopFlatTriangle(this->v1, this->v2, newX);
 	fillBottomFlatTriangle(this->v2, newX, this->v3);
 }
